+++ src/components/super/i-static-page/modules/const.js
@@ -6,6 +6,8 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
+'use strict';
+
 /**
  * Cache of files
  * @type {object}
@@ -22,7 +24,7 @@ exports.folders = Object.create(null);
  * RegExp to determine paths that refer to a folder
  * @type {RegExp}
  */
-exports.isFolder = /[\\/]+$/;
+exports.isFolder = /[/\\]+$/;
 
 /**
  * RegExp to determine URL declaration