+++ src/components/super/i-static-page/modules/ss-helpers/helpers.js
@@ -6,19 +6,27 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
+'use strict';
+
 const
-	{webpack} = require('@config/config');
+	fs = require('node:fs');
+
+const
+	{webpack, src, csp} = require('@config/config'),
+	getHash = include('build/hash');
+
+exports.canLoadStylesDeferred = !webpack.externalizeInline() && !csp.nonce();
 
 exports.needInline = needInline;
 
 /**
- * Returns true if should inline resources
+ * Returns true if the code should be included inline in the tag
  *
- * @param {boolean} [forceInline]
+ * @param {boolean} [forceInline] - if set to true, the function will always return true
  * @returns {boolean}
  */
 function needInline(forceInline) {
-	return Boolean(webpack.fatHTML() || webpack.inlineInitial() || forceInline);
+	return Boolean(forceInline || webpack.fatHTML() || webpack.inlineInitial());
 }
 
 exports.addPublicPath = addPublicPath;
@@ -27,7 +35,7 @@ exports.addPublicPath = addPublicPath;
  * Attaches the `publicPath` property to the specified path
  *
  * @param {(string|Array<string>)} path
- * @returns {(string|Array<String>)}
+ * @returns {(string|Array<string>)}
  */
 function addPublicPath(path) {
 	const
@@ -62,4 +70,38 @@ function addPublicPath(path) {
 
 		return res;
 	}
+}
+
+exports.emitFile = emitFile;
+
+/**
+ * Information about output file paths:
+ *
+ *  1. outputPath - an absolute path to the file
+ *  2. loadPath - a path to load the file, taking into account the configured public path
+ *
+ * @typedef {{
+ *   outputPath: string,
+ *   loadPath: string
+ * }} PathData
+ */
+
+/**
+ * Creates a new file with the given content and stores it in the output directory
+ *
+ * @param {string} content - file content
+ * @param {string} fileName - file name
+ *
+ * @returns {PathData}
+ */
+function emitFile(content, fileName) {
+	fileName = webpack.output({name: fileName, hash: getHash(content)});
+
+	const
+		outputPath = src.clientOutput(fileName),
+		loadPath = webpack.publicPath(fileName);
+
+	fs.writeFileSync(outputPath, content);
+
+	return {outputPath, loadPath};
 }