+++ src/components/super/i-static-page/modules/ss-helpers/page.js
@@ -6,9 +6,17 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
+'use strict';
+
+const {
+	Libs,
+	StyleLibs,
+	Links
+} = require('../interface');
+
 const
 	config = require('@config/config'),
-	{src, webpack, csp} = config;
+	{src, webpack} = config;
 
 const
 	fs = require('fs-extra');
@@ -24,11 +32,10 @@ const
 
 const
 	{getVarsDecl} = include('src/components/super/i-static-page/modules/ss-helpers/base-declarations'),
-	{needInline, addPublicPath} = include('src/components/super/i-static-page/modules/ss-helpers/helpers');
+	{needInline, addPublicPath, canLoadStylesDeferred} = include('src/components/super/i-static-page/modules/ss-helpers/helpers');
 
 const
-	canLoadStylesDeferred = !csp.nonce(),
-	needLoadStylesAsJS = webpack.dynamicPublicPath();
+	needLoadStylesAsJS = Boolean(webpack.dynamicPublicPath());
 
 const defAttrs = {
 	crossorigin: webpack.publicPath() === '' ? undefined : 'anonymous'
@@ -39,16 +46,16 @@ exports.getPageScriptDepsDecl = getPageScriptDepsDecl;
 /**
  * Returns code to load script dependencies of a page.
  *
- * The function returns JS code to load the library by using JS.
+ * The function returns JS code to load the library using JS.
  * You need to put this declaration within a script tag or use the `wrap` option.
  *
  * @param {Array<string>} dependencies - the list of dependencies to load
- * @param {object} opts
- * @param {Object.<string>} opts.assets - a dictionary with static page assets
- * @param {boolean} [opts.wrap] - if true, the final code is wrapped by a script tag
+ * @param {object} [opts] - additional options
+ * @param {Object<string>} opts.assets - a dictionary with static page assets
+ * @param {boolean} [opts.wrap] - if set to true, the final code will be wrapped by a `<script>` tag
  * @returns {string}
  */
-function getPageScriptDepsDecl(dependencies, {assets, wrap} = {}) {
+function getPageScriptDepsDecl(dependencies, {assets, wrap, js} = {}) {
 	if (!dependencies) {
 		return '';
 	}
@@ -58,8 +65,8 @@ function getPageScriptDepsDecl(dependencies, {assets, wrap} = {}) {
 
 	for (const dep of dependencies) {
 		const scripts = [
-			getScriptDeclByName(`${dep}_tpl`, {assets}),
-			getScriptDeclByName(dep, {assets})
+			getScriptDeclByName(`${dep}_tpl`, {assets, js}),
+			getScriptDeclByName(dep, {assets, js})
 		];
 
 		// We can't compile styles into static CSS files because
@@ -87,13 +94,14 @@ exports.getPageStyleDepsDecl = getPageStyleDepsDecl;
 /**
  * Returns code to load style dependencies of a page.
  *
- * The function can return JS code to load the style by using `document.write` or pure CSS to inline.
- * You may use the `wrap` option to wrap the final code with a tag to load.
+ * The function can return JS code to load the style using `document.write` or pure CSS for inline.
+ * You may use the wrap option to wrap the final code with a tag for loading.
  *
  * @param {Array<string>} dependencies - the list of dependencies to load
- * @param {Object.<string>} assets - a dictionary with static page assets
- * @param {boolean} [wrap] - if true, the final code is wrapped by a tag to load
- * @param {boolean} [js] - if true, the function will always return JS code to load the dependency
+ * @param {object} opts - additional options
+ * @param {Object<string>} opts.assets - a dictionary with static page assets
+ * @param {boolean} [opts.wrap] - if set to true, the final code is wrapped by a tag to load
+ * @param {boolean} [opts.js] - if set to true, the function will always return JS code to load the dependency
  * @returns {string}
  */
 function getPageStyleDepsDecl(dependencies, {assets, wrap, js}) {
@@ -119,56 +127,60 @@ function getPageStyleDepsDecl(dependencies, {assets, wrap, js}) {
 exports.getScriptDeclByName = getScriptDeclByName;
 
 /**
- * Returns code to load a script by the specified name.
- * The names are equal with entry points from "src/entries".
+ * Returns code to load a script with the specified name.
+ * The names correspond to entry points from "src/entries".
  *
- * The function returns JS code to load the library by using JS.
+ * The function returns JS code to load the library using JS.
  * You need to put this declaration within a script tag or use the `wrap` option.
  *
  * @param {string} name
- * @param {Object.<string>} assets - a dictionary with static page assets
- * @param {boolean} [optional] - if true, the missing of this script won't throw an error
- * @param {boolean} [defer=true] - if true, the script is loaded with the "defer" attribute
- * @param {boolean} [inline] - if true, the script is placed as a text
- * @param {boolean} [wrap] - if true, the final code is wrapped by a script tag
+ * @param {object} opts - additional options
+ * @param {Object<string>} opts.assets - a dictionary with static page assets
+ * @param {boolean} [opts.optional] - if set to true, the missing of this script won't throw an error
+ * @param {boolean} [opts.defer] - if set to true, the script is loaded with the "defer" attribute
+ * @param {boolean} [opts.inline] - if set to true, the script is placed as a text
+ * @param {boolean} [opts.wrap] - if set to true, the final code will be wrapped by a `<script>` tag
+ * @param {boolean} [opts.js] - if set to true, the function will always return JS code to load the dependency
  * @returns {string}
+ *
+ * @throws {Error} if the dependency with the specified name does not exist
  */
 function getScriptDeclByName(name, {
 	assets,
 	optional,
 	defer = true,
 	inline,
-	wrap
+	wrap,
+	js = false
 }) {
 	let
 		decl;
 
-	if (needInline(inline)) {
-		if (assets[name]) {
-			const
-				filePath = src.clientOutput(assets[name].path);
-
-			if (fs.existsSync(filePath)) {
-				decl = `include('${filePath}');`;
-			}
-
-		} else {
-			if (!optional) {
-				throw new ReferenceError(`A script by the name "${name}" is not defined`);
-			}
-
+	if (!assets[name] && !js) {
+		if (optional) {
 			return '';
+		}
+
+		throw new ReferenceError(`A script with the name "${name}" is not defined`);
+	}
+
+	if (needInline(inline)) {
+		const
+			filePath = src.clientOutput(assets[name].path);
+
+		if (fs.existsSync(filePath)) {
+			decl = `include('${filePath}');`;
 		}
 
 	} else {
 		decl = getScriptDecl({
 			...defAttrs,
 			defer,
-			js: true,
-			src: addPublicPath([`PATH['${name}']`])
+			js,
+			src: js ? addPublicPath([`PATH['${name}']`]) : assets[name].publicPath
 		});
 
-		if (optional) {
+		if (optional && js) {
 			decl = `if ('${name}' in PATH) {
 	${decl}
 }`;
@@ -181,20 +193,23 @@ function getScriptDeclByName(name, {
 exports.getStyleDeclByName = getStyleDeclByName;
 
 /**
- * Returns code to load a style by the specified name.
- * The names are equal with entry points from "src/entries".
+ * Returns code to load a style with the specified name.
+ * The names correspond to entry points from "src/entries".
  *
- * The function can return JS code to load the style by using JS or pure CSS to inline.
- * You may use the `wrap` option to wrap the final code with a tag to load.
+ * The function can return JS code to load the style using JS or pure CSS for inline.
+ * You may use the `wrap` option to wrap the final code with a tag for loading.
  *
  * @param {string} name
- * @param {Object.<string>} assets - a dictionary with static page assets
- * @param {boolean} [optional] - if true, the missing of this style won't throw an error
- * @param {boolean} [defer] - if true, the style is loaded only after loading of the whole page
- * @param {boolean} [inline] - if true, the style is placed as a text
- * @param {boolean} [wrap] - if true, the final code is wrapped by a tag to load
- * @param {boolean} [js] - if true, the function will always return JS code to load the dependency
+ * @param {object} opts - additional options
+ * @param {Object<string>} opts.assets - a dictionary with static page assets
+ * @param {boolean} [opts.optional] - if set to true, the missing of this style won't throw an error
+ * @param {boolean} [opts.defer] - if set to true, the style is loaded only after loading of the whole page
+ * @param {boolean} [opts.inline] - if set to true, the style is placed as a text
+ * @param {boolean} [opts.wrap] - if set to true, the final code is wrapped by a tag to load
+ * @param {boolean} [opts.js] - if set to true, the function will always return JS code to load the dependency
  * @returns {string}
+ *
+ * @throws {Error} if the dependency with the specified name does not exist
  */
 function getStyleDeclByName(name, {
 	assets,
@@ -214,29 +229,33 @@ function getStyleDeclByName(name, {
 	let
 		decl;
 
+	if (!assets[rname] && !js) {
+		if (optional) {
+			return '';
+
+		}
+
+		throw new ReferenceError(`A style with the name "${name}" is not defined`);
+	}
+
 	if (needInline(inline)) {
-		if (assets[rname]) {
-			const
-				filePath = src.clientOutput(assets[rname].path);
+		const
+			filePath = src.clientOutput(assets[rname].path);
 
-			if (fs.existsSync(filePath)) {
-				decl = getStyleDecl({...defAttrs, js}, `include('${filePath}');`);
-			}
-
-		} else if (!optional) {
-			throw new ReferenceError(`A style by the name "${name}" is not defined`);
+		if (fs.existsSync(filePath)) {
+			decl = getStyleDecl({...defAttrs, js}, `include('${filePath}');`);
 		}
 
 	} else {
 		decl = getStyleDecl({
 			...defAttrs,
 			defer,
-			js: true,
+			js,
 			rel: 'stylesheet',
-			src: addPublicPath([`PATH['${rname}']`])
+			src: js ? addPublicPath([`PATH['${rname}']`]) : assets[rname].publicPath
 		});
 
-		if (optional) {
+		if (optional && js) {
 			decl = `if ('${rname}' in PATH) {
 	${decl}
 }`;
@@ -253,18 +272,16 @@ function getStyleDeclByName(name, {
 exports.generateInitJS = generateInitJS;
 
 /**
- * Generates js script to initialize the specified page
+ * Generates a JS script to initialize the specified page
  *
- * @param pageName
- *
- * @param deps - a dictionary with external libraries to load
- * @param ownDeps - the page dependencies
- *
- * @param assets - a dictionary with static page assets
- * @param assetsRequest - should or not do a request for assets.js
- *
- * @param rootAttrs - attributes for the root tag
- *
+ * @param {string} pageName - the page name
+ * @param {object} page - the page parameters
+ * @param {{headScripts: Libs, scripts: Libs, links: Links, styles: StyleLibs}} page.deps - a dictionary
+ *   with external libraries to load
+ * @param {Array<string>} page.ownDeps - the page dependencies
+ * @param {Object<string>} page.assets - a dictionary with static page assets
+ * @param {boolean} page.assetsRequest - should or not do a request for assets.js
+ * @param {object} page.rootAttrs - attributes for the root tag
  * @returns {Promise<void>}
  */
 async function generateInitJS(pageName, {
@@ -314,13 +331,13 @@ async function generateInitJS(pageName, {
 
 	// - block scripts
 	body.push(
-		getScriptDeclByName('std', {assets, optional: true}),
+		getScriptDeclByName('std', {assets, optional: true, js: true}),
 		await loadLibs(deps.scripts, {assets, js: true}),
 
-		getScriptDeclByName('index-core', {assets, optional: true}),
-		getScriptDeclByName('vendor', {assets, optional: true}),
+		getScriptDeclByName('index-core', {assets, optional: true, js: true}),
+		getScriptDeclByName('vendor', {assets, optional: true, js: true}),
 
-		getPageScriptDepsDecl(ownDeps, {assets})
+		getPageScriptDepsDecl(ownDeps, {assets, js: true})
 	);
 
 	const bodyInitializer = `