+++ ssr-example.js
@@ -44,6 +44,6 @@ require('./dist/ssr/p-v4-components-demo')
 	})
 
 	.then((res) => {
-		require('fs').writeFileSync('ssr-example.html', res);
+		require('node:fs').writeFileSync('ssr-example.html', res);
 	});