+++ config/CHANGELOG.md
@@ -9,6 +9,20 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v4.0.0-beta.33 (2023-10-18)
+
+#### :rocket: New Feature
+
+* Added new config option `webpack.externalizeInline`
+
+## v4.0.0-beta.30 (2023-10-11)
+
+#### :rocket: New Feature
+
+* Added default values for `runtime` theme parameters
+* Added a new option `module-parallelism` option, which sets the [`parallelism`](https://webpack.js.org/configuration/other-options/#parallelism) option for a webpack
+* Added a new option `trace-build-times`, which enables the build time tracing
+
 ## v4.0.0-beta.20 (2023-09-13)
 
 #### :rocket: New Feature