+++ src/components/global/g-def/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v4.0.0-beta.30 (2023-10-11)
+
+#### :rocket: New Feature
+
+* Added possibility to configure the theme attribute
+
 ## v4.0.0-beta.27 (2023-09-22)
 
 #### :bug: Bug Fix