+++ src/components/super/i-static-page/CHANGELOG.md
@@ -9,7 +9,19 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
-## v4.0.0-beta.?? (2023-??-??)
+## v4.0.0-beta.33 (2023-10-18)
+
+#### :rocket: New Feature
+
+* Added the ability to build initial HTML file without any inline Javascript by configuring `webpack.externalizeInline`
+
+## v4.0.0-beta.30 (2023-10-11)
+
+#### :bug: Bug Fix
+
+* Fix initializing of the `helpers` and `providers` blocks for SSR
+
+## v4.0.0-beta.29 (2023-10-02)
 
 #### :bug: Bug Fix